+++ SuperBuild/External_CIL.cmake
@@ -1,6 +1,6 @@
 #========================================================================
 # Author: Edoardo Pasca
-# Copyright 2018, 2020 STFC
+# Copyright 2018 UKRI STFC
 #
 # This file is part of the CCP SyneRBI (formerly PETMR) Synergistic Image Reconstruction Framework (SIRF) SuperBuild.
 #
@@ -19,7 +19,7 @@
 #=========================================================================
 
 #This needs to be unique globally
-set(proj CCPi-Framework)
+set(proj CIL)
 
 # Set dependency list
 set(${proj}_DEPENDENCIES "")
@@ -60,8 +60,7 @@ if(NOT ( DEFINED "USE_SYSTEM_${externalProjName}" AND "${USE_SYSTEM_${externalPr
           -DPYTHON_DEST_DIR:PATH=${PYTHON_DEST}
           -DOPENMP_INCLUDES:PATH=${OPENMP_INCLUDES}
           -DCIL_VERSION:STRING=${${proj}_TAG}
-      # INSTALL_COMMAND ${CMAKE_COMMAND} -E env CIL_VERSION=${${proj}_TAG} ${CMAKE_COMMAND} --build . --target install && ${CMAKE_COMMAND} -E copy_directory ${${proj}_SOURCE_DIR}/Wrappers/Python/data ${SUPERBUILD_INSTALL_DIR}/share/ccpi/ && ${PYTHON_EXECUTABLE} ${CMAKE_SOURCE_DIR}/patches/cil-patch.py ${${proj}_SOURCE_DIR}/Wrappers/Python/ccpi/framework/TestData.py ${PYTHON_DEST}/ccpi/framework/TestData.py
-      INSTALL_COMMAND ${CMAKE_COMMAND} --build . --target install && ${CMAKE_COMMAND} -E copy_directory ${${proj}_SOURCE_DIR}/Wrappers/Python/data ${SUPERBUILD_INSTALL_DIR}/share/ccpi/ && ${PYTHON_EXECUTABLE} ${CMAKE_SOURCE_DIR}/patches/cil-patch.py ${${proj}_SOURCE_DIR}/Wrappers/Python/ccpi/utilities/dataexample.py ${PYTHON_DEST}/ccpi/utilities/dataexample.py
+      INSTALL_COMMAND ${CMAKE_COMMAND} --build . --target install && ${CMAKE_COMMAND} -E copy_directory ${${proj}_SOURCE_DIR}/Wrappers/Python/data ${SUPERBUILD_INSTALL_DIR}/share/cil/ && ${PYTHON_EXECUTABLE} ${CMAKE_SOURCE_DIR}/patches/cil-patch.py ${${proj}_SOURCE_DIR}/Wrappers/Python/cil/utilities/dataexample.py ${PYTHON_DEST}/cil/utilities/dataexample.py
       DEPENDS ${${proj}_DEPENDENCIES}
     )
 
@@ -76,18 +75,58 @@ if(NOT ( DEFINED "USE_SYSTEM_${externalProjName}" AND "${USE_SYSTEM_${externalPr
   set(${proj}_INCLUDE_DIR ${${proj}_SOURCE_DIR})
 
   add_test(NAME CIL_FRAMEWORK_TESTS_1
-           COMMAND ${PYTHON_EXECUTABLE} -m unittest discover -p test_Data*.py 
+           COMMAND ${PYTHON_EXECUTABLE} -m unittest discover -p test_DataContainer.py 
            WORKING_DIRECTORY ${${proj}_SOURCE_DIR}/Wrappers/Python/test)
   add_test(NAME CIL_FRAMEWORK_TESTS_2
-           COMMAND ${PYTHON_EXECUTABLE} -m unittest discover -p test_algor*.p 
+           COMMAND ${PYTHON_EXECUTABLE} -m unittest discover -p test_algor*.py 
            WORKING_DIRECTORY ${${proj}_SOURCE_DIR}/Wrappers/Python/test)
   add_test(NAME CIL_FRAMEWORK_TESTS_3
            COMMAND ${PYTHON_EXECUTABLE} -m unittest discover -p test_run_*.py
            WORKING_DIRECTORY ${${proj}_SOURCE_DIR}/Wrappers/Python/test)
-   add_test(NAME CIL_SIRF_TESTS
+  add_test(NAME CIL_SIRF_TESTS
 	   COMMAND ${PYTHON_EXECUTABLE} -m unittest discover -p test_SIRF*.py
            WORKING_DIRECTORY ${${proj}_SOURCE_DIR}/Wrappers/Python/test)
-     
+  add_test(NAME CIL_FRAMEWORK_TESTS_4
+           COMMAND ${PYTHON_EXECUTABLE} -m unittest discover -p test_Block_*.py
+           WORKING_DIRECTORY ${${proj}_SOURCE_DIR}/Wrappers/Python/test)
+  add_test(NAME CIL_FRAMEWORK_TESTS_5
+           COMMAND ${PYTHON_EXECUTABLE} -m unittest discover -p test_dataexample.py
+           WORKING_DIRECTORY ${${proj}_SOURCE_DIR}/Wrappers/Python/test)
+  # disabling this test so that CI builds are happy as in the release version 21.1.0 there is a bug
+  # which is fixed in master. Test to be reinstated after new tag of CIL
+  #add_test(NAME CIL_FRAMEWORK_TESTS_6
+  #         COMMAND ${PYTHON_EXECUTABLE} -m unittest discover -p test_functions.py
+  #         WORKING_DIRECTORY ${${proj}_SOURCE_DIR}/Wrappers/Python/test)
+  add_test(NAME CIL_FRAMEWORK_TESTS_7
+           COMMAND ${PYTHON_EXECUTABLE} -m unittest discover -p test_Gradient.py
+           WORKING_DIRECTORY ${${proj}_SOURCE_DIR}/Wrappers/Python/test)
+  add_test(NAME CIL_FRAMEWORK_TESTS_8
+           COMMAND ${PYTHON_EXECUTABLE} -m unittest discover -p test_NexusReaderWriter.py
+           WORKING_DIRECTORY ${${proj}_SOURCE_DIR}/Wrappers/Python/test)
+  add_test(NAME CIL_FRAMEWORK_TESTS_9
+           COMMAND ${PYTHON_EXECUTABLE} -m unittest discover -p test_Operator.py
+           WORKING_DIRECTORY ${${proj}_SOURCE_DIR}/Wrappers/Python/test)
+  add_test(NAME CIL_FRAMEWORK_TESTS_10
+           COMMAND ${PYTHON_EXECUTABLE} -m unittest discover -p test_PluginsRegularisation.py
+           WORKING_DIRECTORY ${${proj}_SOURCE_DIR}/Wrappers/Python/test)
+  add_test(NAME CIL_FRAMEWORK_TESTS_11
+           COMMAND ${PYTHON_EXECUTABLE} -m unittest discover -p test_quality_measures.py
+           WORKING_DIRECTORY ${${proj}_SOURCE_DIR}/Wrappers/Python/test)
+  add_test(NAME CIL_FRAMEWORK_TESTS_12
+           COMMAND ${PYTHON_EXECUTABLE} -m unittest discover -p test_smoothMixedL21Norm.py
+           WORKING_DIRECTORY ${${proj}_SOURCE_DIR}/Wrappers/Python/test)
+  add_test(NAME CIL_FRAMEWORK_TESTS_13
+           COMMAND ${PYTHON_EXECUTABLE} -m unittest discover -p test_subset.py
+           WORKING_DIRECTORY ${${proj}_SOURCE_DIR}/Wrappers/Python/test)
+  add_test(NAME CIL_FRAMEWORK_TESTS_14
+           COMMAND ${PYTHON_EXECUTABLE} -m unittest discover -p test_SumFunction.py
+           WORKING_DIRECTORY ${${proj}_SOURCE_DIR}/Wrappers/Python/test)
+  add_test(NAME CIL_FRAMEWORK_TESTS_15
+           COMMAND ${PYTHON_EXECUTABLE} -m unittest discover -p test_TranslateFunction.py
+           WORKING_DIRECTORY ${${proj}_SOURCE_DIR}/Wrappers/Python/test)
+  # add_test(NAME CIL_FRAMEWORK_TESTS_ALL
+  #          COMMAND ${PYTHON_EXECUTABLE} -m unittest discover
+  #          WORKING_DIRECTORY ${${proj}_SOURCE_DIR}/Wrappers/Python/test)
   else()
     ExternalProject_Add_Empty(${proj} DEPENDS "${${proj}_DEPENDENCIES}"
       SOURCE_DIR ${${proj}_SOURCE_DIR}