+++ docker/build_gadgetron-ubuntu.sh
@@ -1,11 +1,15 @@
 #!/usr/bin/env bash
 set -ev
-# Gadgetron
+# Gadgetron requirements
 # from https://github.com/gadgetron/gadgetron/blob/master/docker/base/ubuntu_1804/Dockerfile#L8
 apt-get update -qq
 apt-get install -yq --no-install-recommends \
   libhdf5-serial-dev     \
-  libboost-all-dev       \
+  libboost-dev libboost-chrono-dev \
+        libboost-filesystem-dev libboost-thread-dev \
+        libboost-date-time-dev libboost-regex-dev \
+        libboost-program-options-dev libboost-atomic-dev \
+        libboost-test-dev libboost-timer-dev \
   libfftw3-dev           \
   h5utils                \
   jq                     \