+++ docker/user_service-ubuntu.sh
@@ -11,4 +11,22 @@ if [ -f requirements-service.txt ]; then
   pip install -U -r requirements-service.txt
 fi
 
+#install SIRF-Exercises requirements
+cd $INSTALL_DIR/SIRF-Exercises
+if [ -f requirements.txt ]; then
+  conda install -c conda-forge -y --file requirements.txt || \
+  pip install -U -r requirements.txt
+fi
+
+# configure nbstripout
+git config --global filter.nbstripout.extrakeys '
+  metadata.celltoolbar metadata.language_info.codemirror_mode.version
+  metadata.language_info.pygments_lexer metadata.language_info.version'
+  
+nbstripout --install
 # jupyter labextension install @jupyter-widgets/jupyterlab-manager
+
+# CIL-Demos
+git clone https://github.com/TomographicImaging/CIL-Demos.git --recursive -b main $INSTALL_DIR/CIL-Demos
+cd $INSTALL_DIR/CIL-Demos
+nbstripout --install