+++ version_config.cmake
@@ -2,8 +2,8 @@
 # Author: Edoardo Pasca
 # Author: Benjamin A Thomas
 # Author: Kris Thielemans
-# Copyright 2017-2020 University College London
-# Copyright 2017-2020 Science Technology Facilities Council
+# Copyright 2017-2021 University College London
+# Copyright 2017-2021 Science Technology Facilities Council
 #
 # This file is part of the CCP SyneRBI (formerly PETMR) Synergistic Image Reconstruction Framework (SIRF) SuperBuild.
 #
@@ -62,36 +62,20 @@ set(FFTW3double_URL ${FFTW3_URL})
 set(FFTW3double_MD5 ${FFTW3_MD5})
 
 ## HDF5
+set(DEFAULT_HDF5_URL https://github.com/HDFGroup/hdf5/)
 if (WIN32)
+  set(HDF5_REQUIRED_VERSION 1.8.12)
   # 1.8.15 hdf5-targets.cmake refers to non-existent zlib files
   # (or at least this was the case for older Anaconda installations)
-  if (USE_SYSTEM_HDF5)
-    set(HDF5_REQUIRED_VERSION 1.8.12)
-    #set(HDF5_REQUIRED_VERSION 1.8)
-  endif()
-  if (BUILD_MATLAB)
-    set(HDF5_DOWNLOAD_MAJOR_MINOR_VERSION "1.8")
-    set(HDF5_DOWNLOAD_PATCH_VERSION "12")
-  else()
-    set(HDF5_DOWNLOAD_MAJOR_MINOR_VERSION "1.10")
-    set(HDF5_DOWNLOAD_PATCH_VERSION "1")
-  endif()
-  set(HDF5_DOWNLOAD_VERSION "${HDF5_DOWNLOAD_MAJOR_MINOR_VERSION}.${HDF5_DOWNLOAD_PATCH_VERSION}")
-  set(HDF5_URL https://support.hdfgroup.org/ftp/HDF5/releases/hdf5-${HDF5_DOWNLOAD_MAJOR_MINOR_VERSION}/hdf5-${HDF5_DOWNLOAD_VERSION}/src/hdf5-${HDF5_DOWNLOAD_VERSION}.tar.gz)
 else()
   set(HDF5_REQUIRED_VERSION 1.8)
-  set(HDF5_DOWNLOAD_MAJOR_MINOR_VERSION "1.10")
-  set(HDF5_DOWNLOAD_PATCH_VERSION "1")
-  set(HDF5_DOWNLOAD_VERSION "${HDF5_DOWNLOAD_MAJOR_MINOR_VERSION}.${HDF5_DOWNLOAD_PATCH_VERSION}")
-  set(HDF5_URL https://github.com/SyneRBI/assets/releases/download/latest/hdf5-${HDF5_DOWNLOAD_VERSION}.tar.gz)
 endif()
-
-if (${HDF5_DOWNLOAD_VERSION} STREQUAL 1.8.12)
-  set(HDF5_MD5 d804802feb99b87fc668a90e6fa34411 )
-elseif (${HDF5_DOWNLOAD_VERSION} STREQUAL 1.10.1)
-  set(HDF5_MD5 43a2f9466702fb1db31df98ae6677f15 )
+if (BUILD_MATLAB)
+  # Ideally would call MATLAB and use "[majnum,minnum,relnum]=H5.get_libversion()"
+  # but it's been stuck on 1.8.12 for a long time
+  set(DEFAULT_HDF5_TAG hdf5-1_8_12)
 else()
-  message(FATAL_ERROR "Developer error: need to set HDF5_MD5 for version ${HDF5_DOWNLOAD_VERSION}")
+  set(DEFAULT_HDF5_TAG hdf5-1_10_1)
 endif()
 
 ## SWIG
@@ -103,6 +87,7 @@ else(WIN32)
   set(SWIG_URL http://downloads.sourceforge.net/swig/swig-4.0.2.tar.gz )
   set(SWIG_MD5 7c3e46cb5af2b469722cafa0d91e127b )
 endif(WIN32)
+
 option(DEVEL_BUILD "Use current versions of major packages" OFF)
 
 ## Googletest
@@ -147,6 +132,10 @@ set(DEFAULT_TomoPhantom_TAG v1.4)
 set(DEFAULT_NiftyPET_URL https://github.com/pjmark/NIPET )
 set(DEFAULT_NiftyPET_TAG 70b97da0a4eea9445e34831f7393947a37bc77e7)
 
+## parallelproj
+set(DEFAULT_parallelproj_URL https://github.com/gschramm/parallelproj )
+set(DEFAULT_parallelproj_TAG v0.7)
+
 ## SIRF-Contribs
 set(DEFAULT_SIRF-Contribs_URL https://github.com/SyneRBI/SIRF-Contribs )
 set(DEFAULT_SIRF-Contribs_TAG origin/master )
@@ -159,15 +148,12 @@ set(DEFAULT_JSON_URL https://github.com/nlohmann/json.git )
 set(DEFAULT_JSON_TAG v3.9.1)
 
 # CCPi CIL
-set(CIL_VERSION "v20.09")
-set(DEFAULT_CCPi-Framework_URL https://github.com/vais-ral/CCPi-Framework.git)
-set(DEFAULT_CCPi-Framework_TAG "${CIL_VERSION}.1")
-set(DEFAULT_CCPi-Regularisation-Toolkit_URL https://github.com/vais-ral/CCPi-Regularisation-Toolkit.git )
-set(DEFAULT_CCPi-Regularisation-Toolkit_TAG ${CIL_VERSION})
-set(DEFAULT_CCPi-FrameworkPlugins_URL https://github.com/vais-ral/CCPi-FrameworkPlugins.git)
-set(DEFAULT_CCPi-FrameworkPlugins_TAG ${CIL_VERSION})
-set(DEFAULT_CCPi-Astra_URL https://github.com/vais-ral/CCPi-Astra.git)
-set(DEFAULT_CCPi-Astra_TAG "${CIL_VERSION}.1")
+set(DEFAULT_CIL_URL https://github.com/TomographicImaging/CIL.git)
+set(DEFAULT_CIL_TAG "v21.2.0")
+set(DEFAULT_CIL-ASTRA_URL https://github.com/TomographicImaging/CIL-ASTRA.git)
+set(DEFAULT_CIL-ASTRA_TAG "v21.2.0")
+set(DEFAULT_CCPi-Regularisation-Toolkit_URL https://github.com/vais-ral/CCPi-Regularisation-Toolkit.git)
+set(DEFAULT_CCPi-Regularisation-Toolkit_TAG "v20.09")
 
 # range-v3
 set(DEFAULT_range-v3_URL https://github.com/ericniebler/range-v3.git )
@@ -199,12 +185,20 @@ if (DEVEL_BUILD)
   set(DEFAULT_ACE_URL https://github.com/paskino/libace-conda)
   set(DEFAULT_ACE_TAG origin/ACE_version_6.5.9)
 
+  # CCPi CIL
+  set(DEFAULT_CIL_URL https://github.com/TomographicImaging/CIL.git)
+  set(DEFAULT_CIL_TAG origin/master)
+  set(DEFAULT_CIL-ASTRA_URL https://github.com/TomographicImaging/CIL-ASTRA.git)
+  set(DEFAULT_CIL-ASTRA_TAG origin/master)
+  set(DEFAULT_CCPi-Regularisation-Toolkit_URL https://github.com/vais-ral/CCPi-Regularisation-Toolkit.git)
+  set(DEFAULT_CCPi-Regularisation-Toolkit_TAG origin/master)
+
 else()
-  set(DEFAULT_SIRF_TAG v2.2.0)
+  set(DEFAULT_SIRF_TAG v3.1.0)
 
   ## STIR
   set(DEFAULT_STIR_URL https://github.com/UCL/STIR )
-  set(DEFAULT_STIR_TAG release_4)
+  set(DEFAULT_STIR_TAG rel_4.1.1)
 
   ## siemens_to_ismrmrd
   set(DEFAULT_siemens_to_ismrmrd_URL https://github.com/ismrmrd/siemens_to_ismrmrd)
@@ -224,7 +218,7 @@ else()
 endif()
 
 
-# Set the tags for SIRF, STIR, Gadgetron and ISMRMRD
+# Set the tags for SIRF, STIR, Gadgetron and ISMRMRD etc
 # these can be overridden by the user
 SET(SIRF_URL ${DEFAULT_SIRF_URL} CACHE STRING ON)
 SET(SIRF_TAG ${DEFAULT_SIRF_TAG} CACHE STRING ON)
@@ -256,14 +250,14 @@ set(JSON_TAG ${DEFAULT_JSON_TAG} CACHE STRING ON)
 
 set(CCPi-Regularisation-Toolkit_URL ${DEFAULT_CCPi-Regularisation-Toolkit_URL} CACHE STRING ON)
 set(CCPi-Regularisation-Toolkit_TAG ${DEFAULT_CCPi-Regularisation-Toolkit_TAG} CACHE STRING ON)
-set(CCPi-Framework_URL ${DEFAULT_CCPi-Framework_URL} CACHE STRING ON)
-set(CCPi-Framework_TAG ${DEFAULT_CCPi-Framework_TAG} CACHE STRING ON)
-set(CCPi-FrameworkPlugins_URL ${DEFAULT_CCPi-FrameworkPlugins_URL} CACHE STRING ON)
-set(CCPi-FrameworkPlugins_TAG ${DEFAULT_CCPi-FrameworkPlugins_TAG} CACHE STRING ON)
-set(CCPi-Astra_URL ${DEFAULT_CCPi-Astra_URL} CACHE STRING ON)
-set(CCPi-Astra_TAG ${DEFAULT_CCPi-Astra_TAG} CACHE STRING ON)
+set(CIL_URL ${DEFAULT_CIL_URL} CACHE STRING ON)
+set(CIL_TAG ${DEFAULT_CIL_TAG} CACHE STRING ON)
+set(CIL-ASTRA_URL ${DEFAULT_CIL-ASTRA_URL} CACHE STRING ON)
+set(CIL-ASTRA_TAG ${DEFAULT_CIL-ASTRA_TAG} CACHE STRING ON)
 set(astra-toolbox_URL ${DEFAULT_astra-toolbox_URL} CACHE STRING ON)
 set(astra-toolbox_TAG ${DEFAULT_astra-toolbox_TAG} CACHE STRING ON)
+set(astra-python-wrapper_URL ${DEFAULT_astra-toolbox_URL} CACHE STRING ON)
+set(astra-python-wrapper_TAG ${DEFAULT_astra-toolbox_TAG} CACHE STRING ON)
 set(TomoPhantom_URL ${DEFAULT_TomoPhantom_URL} CACHE STRING ON)
 set(TomoPhantom_TAG ${DEFAULT_TomoPhantom_TAG} CACHE STRING ON)
 
@@ -273,8 +267,14 @@ set(NIFTYREG_TAG ${DEFAULT_NIFTYREG_TAG} CACHE STRING ON)
 set(NiftyPET_URL ${DEFAULT_NiftyPET_URL} CACHE STRING ON)
 set(NiftyPET_TAG ${DEFAULT_NiftyPET_TAG} CACHE STRING ON)
 
+set(parallelproj_URL ${DEFAULT_parallelproj_URL} CACHE STRING ON)
+set(parallelproj_TAG ${DEFAULT_parallelproj_TAG} CACHE STRING ON)
+
 set(SIRF-Contribs_URL ${DEFAULT_SIRF-Contribs_URL} CACHE STRING ON)
 set(SIRF-Contribs_TAG ${DEFAULT_SIRF-Contribs_TAG} CACHE STRING ON)
+
+set(HDF5_URL ${DEFAULT_HDF5_URL} CACHE STRING ON)
+set(HDF5_TAG ${DEFAULT_HDF5_TAG} CACHE STRING ON)
 
 set(ITK_URL ${DEFAULT_ITK_URL} CACHE STRING ON)
 set(ITK_TAG ${DEFAULT_ITK_TAG} CACHE STRING ON)
@@ -296,13 +296,18 @@ mark_as_advanced(SIRF_URL SIRF_TAG STIR_URL STIR_TAG
   pet_rd_tools_URL pet_rd_tools_TAG
   glog_URL glog_TAG
   NIFTYREG_URL NIFTYREG_TAG
-  CCPi-Framework_URL CCPi-Framework_TAG
-  CCPi-FrameworkPlugins_URL CCPi-FrameworkPlugins_TAG
+  CIL_URL CIL_TAG
+  CIL-ASTRA_URL CIL-ASTRA_TAG
   CCPi-Regularisation-Toolkit_URL CCPi-Regularisation-Toolkit_TAG
   NiftyPET_URL NiftyPET_TAG
+  parallelproj_URL parallelproj_TAG
   SIRF-Contribs_URL SIRF-Contribs_TAG
   ITK_URL ITK_TAG
   SPM_URL SPM_TAG
   JSON_URL JSON_TAG
   range-v3_URL range-v3_TAG
+  astra-toolbox_URL astra-toolbox_TAG
+  astra-python-wrapper_URL astra-python-wrapper_TAG
+  ACE_URL ACE_TAG
+  
 )